requirements.txt
bzip2==1.0.8
ca-certificates==2023.05.30
click==8.0.4
flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.0.1
jinja2==3.1.2
libffi==3.4.4
markupsafe==2.1.1
ncurses==6.4
numpy==1.25.1
openssl==3.0.10
packaging==23.1
pandas==2.0.3
pdf2image==1.16.3
pillow==10.0.0
pip==23.1.2
pytesseract==0.3.10
python==3.11.4
python-dateutil==2.8.2
pytz==2023.3
readline==8.2
setuptools==67.8.0
six==1.16.0
sqlite==3.41.2
tk==8.6.12
tzdata==2023.3
werkzeug==2.2.3
wheel==0.38.4
xlsxwriter==3.1.2
xz==5.4.2
zlib==1.2.13
